refactor(parser, lexer): extricate lexer code from parser package, TODO: remove all lexer references in AST
