+++ go.work
@@ -1,7 +1,8 @@
-go 1.23.5
+go 1.24.2
 
 use (
 	.
+	./lexer
 	./parser
 	./parser/boolean
 	./repl
