+++ lexer/go.mod
@@ -1,0 +1,3 @@
+module acornlang.dev/lang/lexer
+
+go 1.24.2
