+++ parser/interface.go
@@ -1,5 +1,1 @@
 package parser
-
-type Identifier struct {
-	Name string `@Ident`
-}
